requirements.txt
absl-py==0.15.0
albumentations==0.5.2
annotated-types==0.7.0
astunparse==1.6.3
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clang==5.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==1.6.0
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
filelock==3.16.1
flatbuffers==1.12
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
future==1.0.0
gast==0.4.0
gdown==4.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.17.3
gymnasium==1.1.1
h5py==3.1.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
imgaug==0.4.0
importlib_metadata==8.5.0
Jinja2==3.1.6
joblib==1.4.2
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.6.3
ml-collections==0.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
pandas==1.4.4
Pillow==9.5.0
platformdirs==4.3.6
prettytable==3.11.0
proglog==0.1.12
protobuf==3.19.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyglet==1.5.0
pymunk==5.6.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==0.24.1
scipy==1.10.1
sentry-sdk==2.32.0
setproctitle==1.3.6
shapely==2.0.7
six==1.15.0
smmap==5.0.2
soupsieve==2.7
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.6.0
tensorflow-estimator==2.15.0
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.4.1
torchkit @ git+https://github.com/kevinzakka/torchkit.git@dd5824445b5c3ec9f5b0973c89ffd489500b9eae
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
wandb==0.20.1
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.12.1
x-magical @ git+https://github.com/kevinzakka/x-magical.git@31fa989e2ecd6fcdbcdc6f9b70057ab28f6184f2
zipp==3.20.2
